fixing color of comment
