+++ packages/python-packages/low-level-client-api-gen/low_level_client_api.py
@@ -1,4 +1,4 @@
-from llc_api import console_entry_point
+from llcapi import console_entry_point
 
 if __name__ == "__main__":
     console_entry_point()
